+++ include/bas
@@ -1,0 +1,25 @@
+    bool getFilename (FILE *pFile, char *filename) {
+    
+      int MAXSIZE = 0xFFF;
+      char proclnk[0xFFF];
+      int fno;
+      ssize_t r;
+      if (pFile != NULL)
+      {   
+        fno = fileno(pFile);
+        sprintf(proclnk, "/proc/self/fd/%d", fno);
+        r = readlink(proclnk, filename, MAXSIZE);
+        if (r < 0)
+        {   
+            printf("failed to readlink\n");
+            return 0;
+        }   
+        filename[r] = '\0';
+        printf ("FILENAME: %s\n", filename);
+    
+        return true;
+      }   
+
+      return false;
+    } 
+
